Still more rework

+++ lib/cvx_id.m
@@ -1,5 +1,5 @@
 function y = cvx_id( x ) %#ok
-y = int32(-1);
+y = -1;
 
 % Copyright 2005-2014 CVX Research, Inc.
 % See the file LICENSE.txt for full copyright information.
